Nested projects cause cargo issues
This is sample project generated with `cargo new --bin`. There is a
script which copies Cargo.* and src/ into a nested directory below the
current project. To trigger the issue, run `make_nested.sh`.

I came across this because our project uses Chef's omnibus project to
generate OS packages and this causes a local nested copy to get created.